pyproject.toml
[project]
name = "my-ai-project"
version = "0.1.0"
description = ""
authors = [
    {name = "mangtaeeee",email = "osp9658@naver.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "qdrant-client (>=1.13.3,<2.0.0)",
    "langchain (>=0.3.22,<0.4.0)",
    "openai (>=1.70.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "langchain-openai (>=0.3.11,<0.4.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)"
]

[tool.poetry]
packages = [{include = "my_ai_project", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
